mpd watchdog
